backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
sqlalchemy==2.0.30
pydantic==2.7.1
pillow==12.2.0
numpy==2.4.4
# tensorflow==2.16.1
# opencv-python-headless==4.9.0.80
